[0001] Remove all references to `Response`
All references to responseProducts in any shape or form has been removed
