+++ Example/Trolley/AppDelegate.swift
@@ -16,7 +16,7 @@ public class AppDelegate: UIResponder, UIApplicationDelegate {
 
     public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         
-        DefaultsManager(withKey: "WebSocketQueue").clear()
+//        DefaultsManager(withKey: "WebSocketQueue").clear()
         
         // If you are using our API to store your data you should call this.
         // If not then we would reccomend using `Trolley/Core` in your podfile
@@ -71,7 +71,7 @@ public class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     
     public func applicationDidBecomeActive(_ application: UIApplication) {
-        TRLAnalytics.shared.logSearchQuery("Ripped Jeans")
+//        TRLAnalytics.shared.logSearchQuery("Ripped Jeans")
     }
     
 //
