+++ Trolley/Database/TRLDatabaseCore/Extensions/Core+Exts.swift
@@ -57,60 +57,3 @@ public extension _Basket where P == Product {
     }
 
 }
-
-public extension Product {
-
-    /// To save writing out `TRLNetworkManager.shared`
-    /// everytime. **LAZY IS KEY**
-    private static var networkManager: TRLNetworkManager {
-        return TRLNetworkManager.shared
-    }
-
-    /// <#Description#>
-    ///
-    /// - Returns: <#return value description#>
-    class func getAll() -> ProductsPromise {
-        return self.networkManager.get(.products).responseProducts()
-    }
-
-    /// <#Description#>
-    ///
-    /// - Parameter id: <#id description#>
-    /// - Returns: <#return value description#>
-    class func getProduct(with id: String) -> Promise<Product> {
-        return self.networkManager.get(id, in: .products).responseProduct()
-    }
-
-}
-
-public extension SearchableProducts {
-    /// To save writing out `TRLNetworkManager.shared`
-    /// everytime. **LAZY IS KEY**
-    private static var networkManager: TRLNetworkManager {
-        return TRLNetworkManager.shared
-    }
-    
-    /// - Note:
-    /// This works so you ask the client for all the products and the
-    /// you filter for the ones you want, other wise use the
-    /// `TRLRequest`'s `search(for:)` method if wanting to search for
-    /// items after the user presses search
-    class func getAll() -> Promise<SearchablePromiseReponse> {
-        let promise: Promise<SearchablePromiseReponse>!
-        promise = Promise { fullfill, rejct in
-            self.networkManager
-                .searchForItems(in: .products)
-                .responseJSON(handler: { (json, err) in
-                    if let error = err {
-                        rejct(error)
-                    } else {
-                        let res = SearchablePromiseReponse(json.searchableProducts)
-                        fullfill(res)
-                    }
-                }
-            )
-        }
-        
-        return promise
-    }
-}
